KA-137 - Refactor actionCreators to different files
Simple ones are in the basicActionCreators file. They are in the folder internal along with factories. Everything is exported from the main action creators file - actionCreators.
