+++ onboarding-task/src/actionCreators/internal/addItemFactory.ts
@@ -1,5 +1,5 @@
-import { ADD_ITEM } from '../constants/actionTypes';
-import { IAction } from './IAction';
+import { ADD_ITEM } from '../../constants/actionTypes';
+import { IAction } from '../IAction';
 
 export const addItemFactory = (idGenerator: () => string): ((text: string) => IAction) =>
   (text: string): IAction => ({
