+++ onboarding-task/test/reducerTests/ListItemsReducers/itemsReducers/itemsReducer.test.js
@@ -1,7 +1,7 @@
 import { Map } from 'immutable';
 import { Item } from '../../../../src/models/Item.ts';
 import { deleteItem, enableEditItem, saveChangesToItem, cancelChangesToItem } from '../../../../src/actionCreators/actionCreators.ts';
-import { addItemFactory } from '../../../../src/actionCreators/addItemFactory.ts';
+import { addItemFactory } from '../../../../src/actionCreators/internal/addItemFactory.ts';
 import { itemsReducer } from '../../../../src/reducers/ListItemReducers/itemsReducers/itemsReducer.ts';
 
 describe('Correctly creates reducers', () => {
